Fixes #63
This bug was introduced with CustomFunc. Some generators accept input (we call the input "constraints") and, as of now, there's no way to run a custom generator with default input (the concept doesn't exist yet, defaults are hard coded inside the generators function).

I added a couple of simple integration tests to make sure we don't break it moving forward (to be honest, it's a very simple test but it gives us enough confidence things are glued together correctly for the `-g` flag).

The fact we can't run custom generators using the same function is a known desing limitation of fakedata. And this bug fix is a good reminder we should improve its design in this area.